requirements.txt
opencv-python
google-generativeai
gTTS
requests
pydub
numpy
